Extract deploy/pipeline.go
